rigid bodies working

+++ rigidBody.cpp
@@ -178,10 +178,10 @@ void RigidBody::loadTrimesh(std::string filename){
   triMesh = 
 	std::unique_ptr<btTriangleIndexVertexArray>{
 	new btTriangleIndexVertexArray{
-	  meshIndices.size()/3,
+	  static_cast<int>(meshIndices.size()/3),
 	  meshIndices.data(),
 	  3*sizeof(int),
-	  meshVertices.size()/3,
+	  static_cast<int>(meshVertices.size()/3),
 	  meshVertices.data(),
 	  3*sizeof(typename decltype(meshVertices)::value_type)}};
